+++ tests/testthat/test-quo.R
@@ -210,17 +210,18 @@ test_that("quosures class has subset assign methods", {
 
   x[1:2] <- list(quo(3), quo(4))
   expect_identical(x, quos(3, 4))
-  expect_error(x[2] <- list(4), "Can't assign a double vector to a list of quosures")
+  expect_warning(x[2] <- list(4), "soft-deprecated")
+  ## expect_error(x[2] <- list(4), "Can't assign a double vector to a list of quosures")
 
   x[[2]] <- quo(10)
   expect_identical(x, quos(3, 10))
-  expect_error(x[[2]] <- list(4), "Can't assign a list to a list of quosures")
+  ## expect_error(x[[2]] <- list(4), "Can't assign a list to a list of quosures")
 
   x <- quos(foo = 1, bar = 2)
 
   x$bar <- quo(100)
   expect_identical(x, quos(foo = 1, bar = 100))
-  expect_error(x$foo <- list(4), "Can't assign a list to a list of quosures")
+  ## expect_error(x$foo <- list(4), "Can't assign a list to a list of quosures")
 })
 
 test_that("can remove quosures by assigning NULL", {